+++ akshare/__init__.py
@@ -2090,9 +2090,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.6.69 add: add fund_info_index_em interface
 1.6.70 add: add bond_zh_hs_cov_pre_min interface
 1.6.71 fix: fix bond_zh_hs_cov_pre_min interface
+1.6.72 add: add futures_news_shmet interface
 """
 
-__version__ = "1.6.71"
+__version__ = "1.6.72"
 __author__ = "AKFamily"
 
 import sys